refactor(shred): remove ShredCode wrapper enum
