+++ ledger/src/shred/merkle.rs
@@ -15,7 +15,7 @@ use {
             CodingShredHeader, DataShredHeader, Error, ProcessShredsStats, ShredCommonHeader,
             ShredFlags, ShredVariant, CODING_SHREDS_PER_FEC_BLOCK, DATA_SHREDS_PER_FEC_BLOCK,
             SHREDS_PER_FEC_BLOCK, SIZE_OF_CODING_SHRED_HEADERS, SIZE_OF_DATA_SHRED_HEADERS,
-            SIZE_OF_SIGNATURE,
+            SIZE_OF_NONCE, SIZE_OF_SIGNATURE,
         },
         shredder::ReedSolomonCache,
     },
@@ -247,6 +247,30 @@ impl ShredCode {
         let node = get_merkle_node(shred, SIZE_OF_SIGNATURE..proof_offset).ok()?;
         get_merkle_root(index, node, proof).ok()
     }
+
+    pub(super) fn first_coding_index(&self) -> Option<u32> {
+        let position = u32::from(self.coding_header.position);
+        self.common_header.index.checked_sub(position)
+    }
+
+    pub(super) fn num_data_shreds(&self) -> u16 {
+        self.coding_header.num_data_shreds
+    }
+
+    pub(super) fn num_coding_shreds(&self) -> u16 {
+        self.coding_header.num_coding_shreds
+    }
+
+    pub(super) fn erasure_mismatch(&self, other: &ShredCode) -> bool {
+        let CodingShredHeader {
+            num_data_shreds,
+            num_coding_shreds,
+            position: _,
+        } = &self.coding_header;
+        num_coding_shreds != &other.coding_header.num_coding_shreds
+            || num_data_shreds != &other.coding_header.num_data_shreds
+            || self.first_coding_index() != other.first_coding_index()
+    }
 }
 
 macro_rules! impl_merkle_shred {
@@ -266,7 +290,7 @@ macro_rules! impl_merkle_shred {
         //   ShredCode::capacity(proof_size, chained, resigned).unwrap()
         //       - ShredData::SIZE_OF_HEADERS
         //       + SIZE_OF_SIGNATURE
-        pub(super) fn capacity(proof_size: u8, resigned: bool) -> Result<usize, Error> {
+        pub fn capacity(proof_size: u8, resigned: bool) -> Result<usize, Error> {
             // Merkle proof is generated and signed after coding shreds are
             // generated. Coding shred headers cannot be erasure coded either.
             Self::SIZE_OF_PAYLOAD
@@ -524,7 +548,7 @@ impl<'a> ShredTrait<'a> for ShredCode {
     type SignedData = Hash;
 
     impl_shred_common!();
-    const SIZE_OF_PAYLOAD: usize = shred_code::ShredCode::SIZE_OF_PAYLOAD;
+    const SIZE_OF_PAYLOAD: usize = solana_packet::PACKET_DATA_SIZE - SIZE_OF_NONCE;
     const SIZE_OF_HEADERS: usize = SIZE_OF_CODING_SHRED_HEADERS;
 
     fn from_payload<T>(payload: T) -> Result<Self, Error>
